+++ compiler/evaluator.rs
@@ -1,5 +1,5 @@
 use crate::{builtins, Environment, Object, Objects};
-use slang_parser::{
+use sflynlang_parser::{
     ast::{Expression, Statement},
     Error,
 };
@@ -11,7 +11,10 @@ pub fn evaluate_expression(
     if let Some((identifier, arguments)) = expression.node.get_call() {
         if let Some(identifier_name) = identifier.node.get_identifier() {
             if !environment.get_store().has_key(&identifier_name) {
-                return Err(Error::new_unknown_identifier(identifier.get_position(), identifier_name));
+                return Err(Error::new_unknown_identifier(
+                    identifier.get_position(),
+                    identifier_name,
+                ));
             }
 
             let mut arguments_objects: Vec<Object> = Vec::new();
@@ -23,11 +26,18 @@ pub fn evaluate_expression(
             if identifier_name == "print" {
                 return builtins::print(arguments_objects, identifier.get_position());
             } else if identifier_name == "debug" {
-                return builtins::debug(arguments_objects, identifier.get_position(), environment.is_debug_mode());
+                return builtins::debug(
+                    arguments_objects,
+                    identifier.get_position(),
+                    environment.is_debug_mode(),
+                );
             }
         }
     } else if let Some(string_value) = expression.node.get_string() {
-        return Ok(Object::new(expression.get_position(), Objects::String(string_value)));
+        return Ok(Object::new(
+            expression.get_position(),
+            Objects::String(string_value),
+        ));
     }
 
     Err(Error::new_unknown_token(expression.get_position()))